@@ -1,26 +1,25 @@
 {
+  "private": true,
   "name": "tools",
   "version": "0.0.0",
-  "private": true,
-  "license": "MIT",
-  "type": "module",
   "description": "Free online tools",
+  "license": "MIT",
   "author": "Lai Tsz Hong <info@honghong.me> (https://github.com/tszhong0411/)",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/tszhong0411/tools.git"
   },
+  "type": "module",
   "scripts": {
-    "dev": "next dev",
     "build": "next build",
-    "start": "next start",
+    "dev": "next dev",
+    "format": "prettier -w .",
     "lint": "eslint .",
     "lint:fix": "eslint --fix .",
-    "format": "prettier -w .",
-    "type-check": "tsc --noEmit",
-    "prepare": "husky install"
+    "prepare": "husky install",
+    "start": "next start",
+    "type-check": "tsc --noEmit"
   },
-  "prettier": "@tszhong0411/prettier-config",
   "dependencies": {
     "@monaco-editor/react": "^4.5.1",
     "@radix-ui/react-checkbox": "^1.0.4",
@@ -44,7 +43,7 @@
     "highlight.js": "^11.8.0",
     "lucide-react": "^0.269.0",
     "next": "^13.4.18",
-    "pdfjs-dist": "^3.9.179",
+    "pdfjs-dist": "3.10.111",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-dropzone": "^14.2.3",
@@ -53,16 +52,16 @@
     "react-use": "^17.4.0",
     "sharp": "^0.32.5",
     "tailwind-merge": "^1.14.0",
-    "tailwindcss-animate": "^1.0.6",
+    "tailwindcss-animate": "1.0.7",
     "usehooks-ts": "^2.9.1",
     "uuid": "^9.0.0"
   },
   "devDependencies": {
     "@commitlint/cli": "^17.7.1",
     "@commitlint/config-conventional": "^17.7.0",
-    "@tszhong0411/eslint-config": "^1.0.15",
-    "@tszhong0411/prettier-config": "^1.0.5",
-    "@tszhong0411/tsconfig": "^1.0.1",
+    "@tszhong0411/eslint-config": "1.0.23",
+    "@tszhong0411/prettier-config": "1.0.7",
+    "@tszhong0411/tsconfig": "1.0.3",
     "@types/file-saver": "^2.0.5",
     "@types/node": "20.5.1",
     "@types/react": "18.2.20",
@@ -74,15 +73,7 @@
     "postcss": "^8.4.28",
     "prettier": "^3.0.2",
     "tailwindcss": "^3.3.3",
-    "typescript": "5.1.6"
-  },
-  "lint-staged": {
-    "*.{js,jsx,ts,tsx}": [
-      "eslint"
-    ],
-    "*.{json,css,scss,md,mdx,js,jsx,ts,tsx,cjs,mjs}": [
-      "prettier -c"
-    ]
+    "typescript": "5.2.2"
   },
   "eslintConfig": {
     "extends": [
@@ -93,5 +84,14 @@
       "project": "./tsconfig.json",
       "ecmaVersion": "latest"
     }
-  }
+  },
+  "lint-staged": {
+    "*.{js,jsx,ts,tsx}": [
+      "eslint"
+    ],
+    "*.{json,css,scss,md,mdx,js,jsx,ts,tsx,cjs,mjs}": [
+      "prettier -c"
+    ]
+  },
+  "prettier": "@tszhong0411/prettier-config"
 }
