@@ -2,11 +2,12 @@
 const nextConfig = {
   webpack: (config) => {
     config.externals.push({
-      canvas: 'canvas',
+      canvas: 'canvas'
     })
 
+    // eslint-disable-next-line @typescript-eslint/no-unsafe-return
     return config
-  },
+  }
 }
 
 export default nextConfig
