@@ -1,6 +1,7 @@
+// eslint-disable-next-line import/no-commonjs
 module.exports = {
   plugins: {
     tailwindcss: {},
-    autoprefixer: {},
-  },
+    autoprefixer: {}
+  }
 }
