@@ -1,14 +1,12 @@
 import type { Metadata } from 'next'
 import { Inter } from 'next/font/google'
+
 import '@/styles/globals.css'
-
-import { cn } from '@/lib/utils'
-
 import Footer from '@/components/footer'
 import Header from '@/components/header'
 import CustomToaster from '@/components/toaster'
-
 import { site } from '@/config/site'
+import { cn } from '@/lib/utils'
 
 type RootLayoutProps = {
   children: React.ReactNode
@@ -18,7 +16,7 @@ export const metadata: Metadata = {
   metadataBase: new URL(site.url),
   title: {
     default: site.title,
-    template: `%s ${site.titleTemplate}`,
+    template: `%s ${site.titleTemplate}`
   },
   description: site.description,
   robots: {
@@ -29,15 +27,15 @@ export const metadata: Metadata = {
       follow: true,
       'max-video-preview': -1,
       'max-image-preview': 'large',
-      'max-snippet': -1,
-    },
+      'max-snippet': -1
+    }
   },
   manifest: '/favicon/site.webmanifest',
   twitter: {
     title: site.name,
     card: 'summary_large_image',
     site: '@tszhong0411',
-    creator: '@tszhong0411',
+    creator: '@tszhong0411'
   },
   keywords: site.keywords,
   themeColor: '#000',
@@ -55,9 +53,9 @@ export const metadata: Metadata = {
         width: 1200,
         height: 630,
         alt: site.description,
-        type: 'image/png',
-      },
-    ],
+        type: 'image/png'
+      }
+    ]
   },
   icons: {
     icon: '/favicon/favicon.svg',
@@ -66,17 +64,17 @@ export const metadata: Metadata = {
       {
         url: '/favicon/apple-touch-icon.png',
         sizes: '180x180',
-        type: 'image/png',
-      },
+        type: 'image/png'
+      }
     ],
-    other: [...site.favicons],
-  },
+    other: [...site.favicons]
+  }
 }
 
 const inter = Inter({
   variable: '--font-inter',
   subsets: ['latin'],
-  display: 'swap',
+  display: 'swap'
 })
 
 const RootLayout = (props: RootLayoutProps) => {
