@@ -6,7 +6,7 @@ import { Logo } from './ui/logo'
 
 const Header = () => {
   return (
-    <header className='fixed left-0 right-0 top-0 z-40 bg-black/50 shadow-sm saturate-100 backdrop-blur-[10px]'>
+    <header className='fixed inset-x-0 top-0 z-40 bg-black/50 shadow-sm saturate-100 backdrop-blur-[10px]'>
       <div className='mx-auto flex h-[60px] max-w-4xl items-center justify-between px-8'>
         <Link href='/' aria-label='Home' title='Home'>
           <Logo />
@@ -17,7 +17,6 @@ const Header = () => {
             target='_blank'
             rel='noreferrer noopener'
             aria-label='GitHub'
-            title='GitHub'
           >
             <GithubIcon />
           </a>
