@@ -1,6 +1,6 @@
 export const imageToBase64 = (
   file: Blob,
-  callback: (result: string) => void,
+  callback: (result: string) => void
 ) => {
   const reader = new FileReader()
 
